Use post_thumbnail if it exists
This will grab the post_thumbnail -- if it exists -- for the post image.
If Jetpack & Photon are enabled, it will convert the image to a pretty 600x600 square. Why 600x600? It seemed like a nice size. Facebook provides NO GUIDANCE for what size/shape an image should be. Maybe we should just include the original, raw url?